Separating execution and pre-transform logic
